Corrected font weight descriptions
Corrected variable names from *normal, *bold and *bolder to the correct font names for the weights in the Source Sans typeface i.e. *regular, *semibold and *bold respectively for the 400, 600 and 700 weights. These are the semantically correct names and also helps to avoid confusion if users refer to other examples of Source Sans.